filters: Only fix wildcarded ephemeral port on the first SYN
Set the requiured wildcarded port on the first SYN only, so that any
additional SYNs will not cause mismatches for later packets (e.g.,
FIN).

Fix wildcarded destination IP to the one in the TCP SYN message so
that all flow filters will apply to the same TCP connection
oony. Wildcarded destinations are used with load balancing as the test
can't know which backend is going to be selected.

Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>